requirements.txt
bcrypt==4.0.1
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
PyJWT==2.6.0
python-dotenv==0.21.1
six==1.16.0
SQLAlchemy==2.0.2
typing_extensions==4.4.0
Werkzeug==2.2.2
